refactor: move email-templates plugin to Node app
