+++ imports/node-app/plugins/email-templates/mutations/index.js
@@ -1,0 +1,5 @@
+import renderEmail from "./renderEmail.js";
+
+export default {
+  renderEmail
+};
