+++ imports/node-app/plugins/email-templates/templates/accounts/inviteNewShopMember.js
