+++ imports/node-app/plugins/email-templates/templates/accounts/inviteShopMember.js
