+++ imports/node-app/plugins/email-templates/templates/coreDefault.js
