+++ imports/node-app/registerPlugins.js
@@ -5,6 +5,7 @@ import registerShippingPlugin from "./core-services/shipping/index.js";
 
 /* plugins */
 import registerAddressPlugin from "./plugins/address/index.js";
+import registerEmailTemplatesPlugin from "./plugins/email-templates/index.js";
 import registerShippingRatesPlugin from "./plugins/shipping-rates/index.js";
 import registerSMTPEmailPlugin from "./plugins/email-smtp/index.js";
 import registerSurchargesPlugin from "./plugins/surcharges/index.js";
@@ -23,7 +24,6 @@ import registerDashboardPlugin from "/imports/plugins/core/dashboard/server/no-m
 import registerDiscountCodesPlugin from "/imports/plugins/included/discount-codes/server/no-meteor/register";
 import registerDiscountsPlugin from "/imports/plugins/core/discounts/server/no-meteor/register";
 import registerEmailPlugin from "/imports/plugins/core/email/server/no-meteor/register";
-import registerEmailTemplatesPlugin from "/imports/plugins/included/email-templates/server/register";
 import registerExamplePaymentsPlugin from "/imports/plugins/included/payments-example/server/no-meteor/register";
 import registerFilesPlugin from "/imports/plugins/core/files/server/no-meteor/register";
 import registerI18nPlugin from "/imports/plugins/core/i18n/server/no-meteor/register";
