--- imports/plugins/included/email-templates/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "reaction-plugin-email-templates",
-  "version": "1.0.0",
-  "description": "Default email templates plugin for Reaction",
-  "author": "Reaction Commerce",
-  "license": "GPL-3.0",
-  "dependencies": {
-    "handlebars": "~4.1.2"
-  }
-}
