--- imports/plugins/included/email-templates/server/mutations/index.js
@@ -1,5 +1,0 @@
-import renderEmail from "./renderEmail";
-
-export default {
-  renderEmail
-};
